Created ContainerTestHelper class and moved some common code for ContainerTests
