+++ src/ResourceManagement/AzureBackup/RecoveryServices.Tests/RecoveryServices.Tests.csproj
@@ -75,6 +75,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Helpers\ContainerTestHelpers.cs" />
+    <Compile Include="Helpers\CommonTestHelpers.cs" />
     <Compile Include="Helpers\TestConstants.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ScenarioTests\AzureIaasVMTests\ContainerTests.cs" />
